requirements.txt
streamlit==1.22.0
openai==0.8.0
python-dotenv==0.21.0